+++ js/notice_manager_panel.js
@@ -5,6 +5,23 @@
 var NoticeManager = (function ($, document) {
 	let options = window.notice_manager_options;
 
+	let selectors_notice = [
+		"div.notice",
+		"div.updated",
+	];
+
+	let selectors_warning = [
+		"div.notice-warning",
+		"div.update-nag",
+	];
+
+	let selectors_error = [
+		"div.error",
+		"div.notice-error",
+	];
+
+	let selectors_all = selectors_notice.concat(selectors_warning, selectors_error);
+
 	// wait function used with autoCollapse
 	let wait = function (ms) {
 		var dfd = $.Deferred();
@@ -40,10 +57,11 @@ var NoticeManager = (function ($, document) {
 	});
 
 	// scroll page to top when closing notice panel
-	// cannot convert to arrow function - uses this
-	// could use event.target instead
-	button.on("click", function () {
-		if ($(this).hasClass("screen-meta-active")) {
+	// function used to work with $(this)
+	// using e.target instead
+	// not sure if this should perhaps be e.currentTarget
+	button.on("click", (e) => {
+		if ($(e.target).hasClass("screen-meta-active")) {
 			if (haveClosed) {
 				NoticeManager.addCounter();
 			}
@@ -68,9 +86,9 @@ var NoticeManager = (function ($, document) {
 		console.log(options);
 
 		// bootstrap notices
-		// get all notices that are not explicily marked as `.inline` or `.below-h2`
+		// get all notices that are not explicitly marked as `.inline` or `.below-h2`
 		// we add .update-nag.inline for WordPress Update notice
-		notices = $("div.updated, div.error, div.notice")
+		notices = $( selectors_all.join(', ') )
 			.not(".inline, .below-h2")
 			.add("div.update-nag");
 
@@ -133,9 +151,9 @@ var NoticeManager = (function ($, document) {
 		getNotices: () => notices,
 
 		getNoticesTopPriority: () => {
-			if ( notices.filter('.error').length )
+			if (notices.filter(":visible").filter(selectors_error.join(", ")).length)
 				return 'error';
-			if ( notices.filter('.notice-warning, .update-nag').length )
+			if (notices.filter(":visible").filter(selectors_warning.join(", ")).length)
 				return 'warning';
 			return 'notice';
 		},
@@ -148,7 +166,7 @@ var NoticeManager = (function ($, document) {
 			notices.appendTo(".notice_container").eq(0);
 			$(".notice_container").removeClass("empty"); // .empty removes padding
 
-			haveClosed = true; // initial collection has occured.
+			haveClosed = true; // initial collection has occurred.
 
 			/**
 			 * When dismissible notices are dismissed, check if any notices are left on page.
@@ -156,10 +174,10 @@ var NoticeManager = (function ($, document) {
 			 */
 			$(document).on(
 				"DOMNodeRemoved",
-				"#meta-link-notices-wrap .notice.is-dismissible",
-				(e) => {
+				"#meta-link-notices-wrap div.notice",
+				() => {
 					notices = panel
-						.find("div.updated, div.error, div.notice, div.update-nag")
+						.find(selectors_all.join(", "))
 						.filter(":visible");
 					NoticeManager.maybeRemoveNoticesPanel();
 				}
@@ -169,7 +187,7 @@ var NoticeManager = (function ($, document) {
 		addCounter: () => {
 			if (!button.children('.plugin-count').length){
 				button.append(
-					$("<span/>").text(notices.length).attr({
+					$("<span/>").text(notices.filter(":visible").length).attr({
 						class: "plugin-count",
 					}).addClass(NoticeManager.getNoticesTopPriority())
 				);