+++ notice-manager.php
@@ -2,17 +2,20 @@
 /**
  * Plugin Name: Notice Manager
  * Description: Manage notices on WordPress admin pages. Adds 'Notices' screen-meta-link.
- * Version: 0.14
+ * Version: 0.15
  * Author: abuyoyo
  * Author URI: https://github.com/abuyoyo/
  * Plugin URI: https://github.com/abuyoyo/notice-manager
+ * Update URI: https://github.com/abuyoyo/notice-manager
  */
 defined( 'ABSPATH' ) || die( 'No soup for you!' );
 
 use WPHelper\PluginCore;
 
+require_once 'vendor/autoload.php';
+
 /**
- * Print setting page
+ * Bootstrap plugin and admin page (Tools > Notice Manager)
  */
 new PluginCore(
 	__FILE__,
@@ -27,32 +30,32 @@ new PluginCore(
 				'sections' => [
 					[
 						'id'          => 'notice_manager',
-						// 'title'       => 'N',
-						'description' => 'Setup How notice manager functions.',
+						'description' => 'Setup Notice Manager options.',
+						'description_container' => 'notice-info',
 						'fields'      => [
 							[
 								'id' => 'above_title',
 								'title' => 'Above Title',
 								'type' => 'checkbox',
-								'description' => 'Move all notices above title. (WordPress core moves notices below title using script.)',
+								'description' => 'Simply move all notices above title. WordPress core moves notices below title using script. This script moves them back over the title. This option does not move notices into panel.',
 							],
 							[
 								'id' => 'screen_panel',
 								'title' => 'Notices Panel',
 								'type' => 'checkbox',
-								'description' => 'Enable\disable screen-meta-links \'Notices\' panel.',
+								'description' => 'Enable Screen Meta \'Notices\' panel. User can collect notices into collapsible panel.',
 							],
 							[
 								'id' => 'auto_collect',
 								'title' => 'Auto-Collect Notices',
 								'type' => 'checkbox',
-								'description' => 'Automatic collection of notices into panel.',
+								'description' => 'If Notices panel is enabled - auto-collect notices into panel on page load.',
 							],
 							[
 								'id' => 'auto_collapse',
-								'title' => 'Auto-collapse Panel',
+								'title' => 'Auto-Collapse Panel',
 								'type' => 'checkbox',
-								'description' => 'Notices panel will stay open for a few seconds on page load, and then close automatically.',
+								'description' => 'If auto-collect is enabled - Notices panel will stay open for a few seconds on page load, and then close automatically. Panel will not auto-collapse if it contains `error` level notices.',
 							],
 						]
 					]
@@ -65,6 +68,4 @@ new PluginCore(
 require_once 'src/NoticeManager.php';
 
 
-add_action('plugins_loaded', function(){
-	new NoticeManager();
-});
+add_action( 'plugins_loaded', fn() => new NoticeManager() );